# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Price Tag API",
    "description": "LLM-friendly REST API for creating, managing, and generating professional price tags with PDF export. This API is optimized for function calling by AI assistants and LLMs.",
    "version": "1.0.0",
    "contact": {
      "name": "Price Tag API Support",
      "url": "https://t.me/PriceTagPrinterBot"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://your-domain.com/api",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/generate-pdf-v2": {
      "get": {
        "operationId": "getPdfGenerationInfo",
        "summary": "Get PDF generation capabilities",
        "description": "Returns supported formats, fonts, design types, and limits for PDF generation. Use this to understand available options before generating PDFs.",
        "tags": ["Generation"],
        "responses": {
          "200": {
            "description": "PDF generation information",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PdfGenerationInfoResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "generatePdf",
        "summary": "Generate PDF from price tags",
        "description": "Create a high-quality PDF document containing formatted price tags with customizable themes, fonts, and layouts. Perfect for printing professional price tags.",
        "tags": ["Generation"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GeneratePdfRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "PDF file generated successfully",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            },
            "headers": {
              "Content-Disposition": {
                "schema": {
                  "type": "string"
                },
                "example": "attachment; filename=price-tags-1642781234567.pdf"
              },
              "Content-Length": {
                "schema": {
                  "type": "integer"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    },
    "/generate-html": {
      "post": {
        "operationId": "generateHtml",
        "summary": "Generate HTML preview of price tags",
        "description": "Create an HTML preview of price tags with full styling and print-ready CSS. Use this to preview how price tags will look before generating PDF.",
        "tags": ["Generation"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateHtmlRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "HTML generated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HtmlGenerationResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "PriceTag": {
        "type": "object",
        "required": ["id", "data", "price", "discountPrice"],
        "properties": {
          "id": {
            "type": "integer",
            "description": "Unique identifier for the price tag"
          },
          "data": {
            "oneOf": [
              {"type": "string"},
              {"type": "number"}
            ],
            "description": "Product name or identifier"
          },
          "price": {
            "type": "integer",
            "minimum": 1,
            "description": "Main price in smallest currency unit (e.g., cents)"
          },
          "discountPrice": {
            "type": "integer",
            "minimum": 0,
            "description": "Calculated discount price in smallest currency unit"
          },
          "designType": {
            "type": "string",
            "enum": ["default", "new", "sale", "white", "black", "sunset", "ocean", "forest", "royal", "vintage", "neon", "monochrome", "silver", "charcoal", "paper", "ink", "snow"],
            "description": "Theme type for visual design"
          },
          "hasDiscount": {
            "type": "boolean",
            "description": "Whether discount pricing should be displayed"
          },
          "priceFor2": {
            "type": "integer",
            "minimum": 1,
            "description": "Special price when buying 2 items"
          },
          "priceFrom3": {
            "type": "integer",
            "minimum": 1,
            "description": "Special price when buying 3 or more items"
          }
        }
      },
...
      "GeneratePdfRequest": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PriceTag"
            },
            "minItems": 1,
            "maxItems": 1000,
            "description": "Price tags to include in PDF"
          },
          "settings": {
            "$ref": "#/components/schemas/PriceTagSettings"
          },
          "format": {
            "type": "string",
            "enum": ["A4", "A3", "Letter"],
            "default": "A4",
            "description": "PDF page format"
          },
          "margin": {
            "$ref": "#/components/schemas/PdfMargin"
          }
        }
      },
      "GenerateHtmlRequest": {
        "type": "object",
        "required": ["items"],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PriceTag"
            },
            "minItems": 1,
            "maxItems": 1000,
            "description": "Price tags to include in HTML"
          },
          "settings": {
            "$ref": "#/components/schemas/PriceTagSettings"
          }
        }
      },
      "PriceTagSettings": {
        "type": "object",
        "properties": {
          "design": {
            "type": "boolean",
            "default": false,
            "description": "Enable discount display globally"
          },
          "designType": {
            "type": "string",
            "enum": ["default", "new", "sale", "white", "black", "sunset", "ocean", "forest", "royal", "vintage", "neon", "monochrome", "silver", "charcoal", "paper", "ink", "snow"],
            "default": "default",
            "description": "Global design theme type"
          },
          "currentFont": {
            "type": "string",
            "enum": ["montserrat", "nunito", "inter", "mont"],
            "default": "montserrat",
            "description": "Font family for price tags"
          },
          "discountAmount": {
            "type": "integer",
            "minimum": 0,
            "default": 500,
            "description": "Fixed discount amount in smallest currency unit"
          },
          "maxDiscountPercent": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "default": 5,
            "description": "Maximum discount percentage allowed"
          },
          "showThemeLabels": {
            "type": "boolean",
            "default": true,
            "description": "Show NEW/SALE labels on themed price tags"
          },
          "discountText": {
            "type": "string",
            "default": "Special Offer",
            "description": "Text to display for discount information"
          },
          "cuttingLineColor": {
            "type": "string",
            "pattern": "^#[0-9a-fA-F]{6}$",
            "default": "#cccccc",
            "description": "Color for cutting guidelines (hex format)"
          }
        }
      },
      "PdfMargin": {
        "type": "object",
        "properties": {
          "top": {
            "type": "string",
            "default": "0",
            "description": "Top margin (e.g., '10mm', '0.5in')"
          },
          "right": {
            "type": "string",
            "default": "0",
            "description": "Right margin"
          },
          "bottom": {
            "type": "string",
            "default": "0",
            "description": "Bottom margin"
          },
          "left": {
            "type": "string",
            "default": "0",
            "description": "Left margin"
          }
        }
      },
      "ApiResponse": {
        "type": "object",
        "required": ["success"],
        "properties": {
          "success": {
            "type": "boolean",
            "description": "Whether the operation was successful"
          },
          "data": {
            "description": "Response data (present when success is true)"
          },
          "error": {
            "type": "string",
            "description": "Error message (present when success is false)"
          },
          "message": {
            "type": "string",
            "description": "Optional additional information"
          }
        }
      },
...
      "HtmlGenerationResponse": {
        "allOf": [
          {"$ref": "#/components/schemas/ApiResponse"},
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "html": {"type": "string"},
                  "itemCount": {"type": "integer"}
                }
              }
            }
          }
        ]
      },
      "PdfGenerationInfoResponse": {
        "allOf": [
          {"$ref": "#/components/schemas/ApiResponse"},
          {
            "type": "object",
            "properties": {
              "data": {
                "type": "object",
                "properties": {
                  "supportedFormats": {"type": "array", "items": {"type": "string"}},
                  "supportedFonts": {"type": "array", "items": {"type": "string"}},
                  "supportedDesignTypes": {"type": "array", "items": {"type": "string"}},
                  "maxItemsPerRequest": {"type": "integer"},
                  "version": {"type": "string"}
                }
              }
            }
          }
        ]
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Validation error occurred",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiResponse"
            },
            "example": {
              "success": false,
              "error": "Price is required and must be a positive number"
            }
          }
        }
      },
...
    }
  },
...
}